requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
aiofiles==23.2.1
PyYAML==6.0.1
